# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Cloud Aegis API
  version: 1.0.0-beta
  description: |
    Multi-cloud CSPM platform API providing security findings, compliance posture,
    attack path analysis, remediation workflows, cost management, and integrated
    terminal access. RBAC is enforced via JWT claims with four roles:
    viewer, requester, operator, admin (least to most privileged).
  contact:
    name: Liem Vo-Nguyen
    email: liem@vonguyen.io
  license:
    name: MIT
    identifier: MIT
servers:
  - url: https://api.cloudforge-demo.lvonguyen.com
    description: Production (Fly.io)
  - url: http://localhost:8080
    description: Development
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/costs/estimate:
    get:
      operationId: getCostEstimate
      summary: Cost estimate for a resource
      description: "RBAC: viewer, operator, admin. Returns estimated monthly cost for a given resource type, provider, and size."
      tags: [Costs]
      x-rbac: [viewer, operator, admin]
      parameters:
        - name: resource_type
          in: query
          required: true
          schema:
            type: string
        - name: provider
          in: query
          required: true
          schema:
            type: string
        - name: size
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cost estimate
          content:
            application/json:
              schema:
                type: object
                properties:
                  resource_type:
                    type: string
                  provider:
                    type: string
                  size:
                    type: string
                  monthly_cost_usd:
                    type: number
                  description:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Unknown resource type or size
...
  /api/v1/ai/nlq:
    post:
      operationId: queryNLQ
      summary: Natural language query to structured filters
      description: "RBAC: operator, admin. Scope-guarded. Rate-limited to ~20 req/min per user."
      tags: [AI/NLQ]
      x-rbac: [operator, admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
                  maxLength: 500
      responses:
        "200":
          description: Parsed structured filters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NLQResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          description: Rate limit exceeded
        "503":
          description: AI provider not configured
  /api/v1/ai/usage:
    get:
      operationId: getAIUsage
      summary: AI budget and usage status
      description: "RBAC: admin only."
      tags: [AI/NLQ]
      x-rbac: [admin]
      responses:
        "200":
          description: AI usage and budget status
          content:
            application/json:
              schema:
                type: object
                properties:
                  monthly_budget_cents:
                    type: integer
                  spent_cents:
                    type: integer
                  remaining_cents:
                    type: integer
                  exhausted:
                    type: boolean
                  tiers:
                    type: array
                    items:
                      type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ---------------------------------------------------------------------------
  # Deploy
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT with claims: sub, email, groups (aegis-admin, aegis-operator,
        aegis-requester, aegis-viewer), scope, tenant_id.
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Insufficient permissions or out of scope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    NLQResponse:
      type: object
      properties:
        severity:
          type: array
          items:
            type: string
        provider:
          type: array
          items:
            type: string
        category:
          type: array
          items:
            type: string
        status:
          type: array
          items:
            type: string
        environment:
          type: array
          items:
            type: string
        text:
          type: string
...